requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
Flask-Login
eventlet
Werkzeug
pillow
flask_wtf
email_validator
gunicorn
